XMLHttpRequest/send-authentication-existing-session-interactive.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - "Basic" authenticated requests with user name and password from interactive session</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" data-tested-assertations="following::code[contains(@title,'http-authorization')]/.." />
  </head>
  <body>
    <p>Please follow these steps to complete the test:</p>
    <ol>
      <li>Load <a href="resources/auth3/auth.php">page</a> and authenticate with "user" and "pass"</li>
      <li>Go back</li>
      <li>Click <a href="?dotest">complete test</a></li>
    </ol>
    <div id="log"></div>
    <script>
    if(location.search === '?dotest'){
      test(function() {
        var client = new XMLHttpRequest(),
          urlstart = location.host + location.pathname.replace(/\/[^\/]*$/, '/')
        client.open("GET", location.protocol+'//'+urlstart + "resources/auth3/auth.php", false)
        client.send(null)
        assert_equals(client.responseText, 'user' + "\n" + 'pass')
        assert_equals(client.getResponseHeader('x-challenge'), 'DID-NOT')
      }, document.title)
    }
    </script>
  </body>
</html>
